requirements-prod.txt
asn1crypto==0.24.0
boto==2.48.0
certifi==2018.4.16
cffi==1.14.6
chardet==3.0.4
cryptography==2.2.2
decorator==4.3.0
Django==2.2.0
    django-admin-sortable2==1.0.2
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
    django-bootstrap4==2.0.0
django-compressor==2.4.1
django-debug-toolbar==1.11
django-extensions==2.0.7
    django-filebrowser-no-grappelli==3.8.0
    # django-jet==1.0.7
django-model-utils==3.1.2
django-rest-multiple-models==2.1.0
django-simple-history==2.1.1
django-tinymce4-lite==1.7.1
djangorestframework==3.8.2
docutils==0.14
GDAL==3.2.1
googlemaps==4.0.0
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.20.1
idna==2.6
ipython-genutils==0.2.0
Jinja2==2.10
jmespath==0.9.3
# jsmin==2.2.2
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
nbformat==4.4.0
numpy==1.21.2
oauthlib==3.0.0
OWSLib
Pillow
plotly
psycopg2==2.9.1
pycparser==2.18
pycrypto==2.6.1
pyenchant==2.0.0
Pygments==2.2.0
PyJWT==2.2.0
PyPDF2==1.26.0
pyproj==2.0
python-dateutil==2.8.2
python-decouple==3.1
python-dotenv==0.19.1
pytz==2018.4
pyxero==0.9.3
PyYAML==3.12
rcssmin==1.0.6
reportlab==3.4.0
requests==2.20.0
requests-oauthlib==0.8.0
rjsmin==1.1.0
Shapely==1.6.1
six==1.16.0
sqlparse==0.4.2
stripe==1.82.1
traitlets==5.1.0
urllib3==1.22
virtualenv==16.0.0
webencodings==0.5.1
xhtml2pdf==0.2.2
